Move TurnService implementation for Coturn to another module
